requirements.txt
botbuilder-core>=4.7.0
asyncio>=3.4
aiohttp==1.3.5
pandas==1.1.3
azureml-sdk==1.19.0
numpy==1.19.2